15th sprint was added: - packages - playerRepository interface and his implementation - removed activity-files into ui package
