+++ app/src/main/java/com/example/playlistmaker/HistoryPrefs.kt
@@ -1,6 +1,7 @@
 package com.example.playlistmaker
 
 import android.content.Context
+import com.example.playlistmaker.domain.models.Track
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
