+++ app/src/main/java/com/example/playlistmaker/ItunesTrackResponse.kt
@@ -1,5 +1,7 @@
 package com.example.playlistmaker
 
+import com.example.playlistmaker.domain.models.Track
+
 class ItunesTrackResponse(
     val resultsCount: Int,
     val results: List<Track>
