+++ app/src/main/java/com/example/playlistmaker/OnClickListenerItem.kt
@@ -1,5 +1,7 @@
 package com.example.playlistmaker
 
+import com.example.playlistmaker.domain.models.Track
+
 
 interface OnClickListenerItem {
     fun onItemClick(track: Track)
