+++ app/src/main/java/com/example/playlistmaker/TrackAdapter.kt
@@ -3,6 +3,7 @@ package com.example.playlistmaker
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import com.example.playlistmaker.domain.models.Track
 
 class TrackAdapter(
     var tracks: ArrayList<Track>,
