+++ app/src/main/java/com/example/playlistmaker/domain/models/Track.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.domain.models
 
 
 import android.os.Parcel
