+++ app/src/main/java/com/example/playlistmaker/domain/repository/PlayerRepository.kt
@@ -1,0 +1,17 @@
+package com.example.playlistmaker.domain.repository
+
+import android.widget.ImageButton
+import android.widget.TextView
+
+interface PlayerRepository {
+
+    fun startPlayer(playBtn: ImageButton)
+
+    fun pausePlayer(playBtn: ImageButton)
+
+    fun preparePlayer(playBtn: ImageButton, timePlaying: TextView, recordsUrl: String?)
+
+    fun playBackControl(playBtn: ImageButton, timePlaying: TextView)
+
+    fun createPlayingTimer(startTime: Long, timePlaying: TextView): Runnable
+}
