+++ app/src/main/java/com/example/playlistmaker/presentation/ui/AudioPlayerActivity.kt
@@ -1,10 +1,7 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.presentation.ui
 
 import android.content.Context
-import android.media.MediaPlayer
 import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
 import android.util.TypedValue
 import android.view.View
 import android.widget.ImageButton
@@ -13,19 +10,16 @@ import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.example.playlistmaker.SearchActivity.Companion.CLICKED_ITEM
+import com.example.playlistmaker.R
+import com.example.playlistmaker.presentation.ui.SearchActivity.Companion.CLICKED_ITEM
+import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.data.repository.PlayerRepositoryImpl
 import java.text.SimpleDateFormat
 import java.util.Locale
 
 class AudioPlayerActivity : AppCompatActivity() {
-    companion object {
-        private const val STATE_DEFAULT = 0
-        private const val STATE_PREPARED = 1
-        private const val STATE_PLAYING = 2
-        private const val STATE_PAUSED = 3
-        private const val UPDATING_DELAY = 400L
-    }
 
+    private val playerRepository by lazy { PlayerRepositoryImpl(context = applicationContext) }
     private lateinit var coverTrack: ImageView
     private lateinit var nameTrack: TextView
     private lateinit var artistsTrack: TextView
@@ -40,9 +34,6 @@ class AudioPlayerActivity : AppCompatActivity() {
     private lateinit var playBtn: ImageButton
     private var recordsUrl: String? = null
     private val dateFormat by lazy { SimpleDateFormat("mm:ss", Locale.getDefault()) }
-    private var mediaPlayer = MediaPlayer()
-    private var playerState = STATE_DEFAULT
-    private var handler = Handler(Looper.getMainLooper())
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_audio_player)
@@ -52,10 +43,10 @@ class AudioPlayerActivity : AppCompatActivity() {
         val imageArrow = findViewById<ImageButton>(R.id.ib_arrowBack)
         init()
         bind(selectedTrack)
-        preparePlayer()
+        playerRepository.preparePlayer(playBtn, timePlaying, recordsUrl)
 
         playBtn.setOnClickListener {
-            playbackControl()
+            playerRepository.playBackControl(playBtn, timePlaying)
         }
 
         imageArrow.setOnClickListener { finish() }
@@ -64,13 +55,13 @@ class AudioPlayerActivity : AppCompatActivity() {
 
     override fun onPause() {
         super.onPause()
-        pausePlayer()
+        playerRepository.pausePlayer(playBtn)
     }
 
     override fun onDestroy() {
         super.onDestroy()
-        mediaPlayer.release()
-        handler.removeCallbacksAndMessages(null)
+        playerRepository.mediaPlayer.release()
+        playerRepository.handler.removeCallbacksAndMessages(null)
     }
 
     private fun bind(track: Track?) {
@@ -128,64 +119,4 @@ class AudioPlayerActivity : AppCompatActivity() {
         playBtn = findViewById(R.id.btn_play)
     }
 
-    private fun preparePlayer() {
-        mediaPlayer.setDataSource(recordsUrl)
-        mediaPlayer.prepareAsync()
-        mediaPlayer.setOnPreparedListener {
-            playBtn.isEnabled = true
-            playerState = STATE_PREPARED
-        }
-        mediaPlayer.setOnCompletionListener {
-            playerState = STATE_PREPARED
-            timePlaying.text = getString(R.string.zero)
-            playBtn.setImageResource(R.drawable.play_vector)
-            handler.removeCallbacksAndMessages(null)
-        }
-    }
-
-    private fun startPlayer() {
-        mediaPlayer.start()
-        playBtn.setImageResource(R.drawable.pause_vector)
-        playerState = STATE_PLAYING
-    }
-
-    private fun pausePlayer() {
-        mediaPlayer.pause()
-        playBtn.setImageResource(R.drawable.play_vector)
-        playerState = STATE_PAUSED
-
-    }
-
-    private fun playbackControl() {
-        when (playerState) {
-            STATE_PLAYING -> {
-                pausePlayer()
-                handler.removeCallbacksAndMessages(null)
-            }
-
-            STATE_PREPARED, STATE_PAUSED -> {
-
-                startPlayer()
-                val startPlaying = System.currentTimeMillis()   //время начала отсчета
-                handler.post(createPlayingTimer(startPlaying))
-            }
-        }
-    }
-
-    private fun createPlayingTimer(startTime: Long): Runnable {
-        return object : Runnable {
-            override fun run() {
-
-                val countedTime = startTime + mediaPlayer.currentPosition
-
-                if (countedTime > 0) {
-                    if (playerState == STATE_PLAYING) {
-                        timePlaying.text = dateFormat.format(mediaPlayer.currentPosition)
-                        handler.postDelayed(this, UPDATING_DELAY)
-                    }
-                }
-
-            }
-        }
-    }
 }
