+++ app/src/main/java/com/example/playlistmaker/presentation/ui/SettingsActivity.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.presentation.ui
 
 import android.content.Intent
 import android.content.SharedPreferences
@@ -7,7 +7,9 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.ImageView
 import android.widget.TextView
+import com.example.playlistmaker.App
 import com.example.playlistmaker.HistoryPrefs.Companion.SHARED_PREFERENCES_HISTORY
+import com.example.playlistmaker.R
 import com.google.android.material.switchmaterial.SwitchMaterial
 
 class SettingsActivity : AppCompatActivity() {
