+++ app/src/main/res/layout/activity_media.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MediaActivity">
+    tools:context=".presentation.ui.MediaActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
