+++ app/src/main/res/layout/activity_search.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="?attr/colorSecondary"
     android:orientation="vertical"
-    tools:context=".SearchActivity">
+    tools:context=".presentation.ui.SearchActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
